+++ ucblock_solver/ucblock_solver.cpp
@@ -91,7 +91,24 @@ int main( int argc , char ** argv )
  // note that the local options are inserted right before the last (nullptr)
  // record in long_opts
 
- docopt_desc = "SMS++ UCBlock solver\n";
+ docopt_desc =
+  "SMS++ UCBlock solver: loads a Unit Commitment problem (a UCBlock) and\n"
+  "solves it with the Solvers of its BlockSolverConfig.\n";
+ docopt_args =
+  "  <file>    SMS++ netCDF file (.nc4) holding a UCBlock: a Block file,\n"
+  "            or a problem file of which only the Block is used;\n"
+  "            <file>[i] selects the i-th Block (or problem) of the file\n";
+ docopt_examples =
+  "  ucblock_solver -B InnerBCfg.txt instance.nc4\n"
+  "      solve with the default configuration, InnerBCfg.txt choosing the\n"
+  "      formulation of the units and of the network\n"
+  "  ucblock_solver instance.nc4\n"
+  "      the same, with the formulation stored in instance.nc4\n"
+  "  ucblock_solver -B InnerBCfg.txt -O sol.nc4 -C OSolCfg.txt instance.nc4\n"
+  "      also save the Solution of the units and of the network in sol.nc4\n"
+  "  ucblock_solver -c myconfig/ instance.nc4\n"
+  "      use the Configuration files in myconfig/, e.g. a modified copy\n"
+  "      of the installed ones\n";
  /*short_opts.append( my_short_opts );
  long_opts.insert( std::prev( long_opts.end() ) ,
 		   my_long_opts.begin() , my_long_opts.end() );